pyproject.toml
[tool.poetry]
name = "buzzline"
version = "0.1.0"
description = "Buzzline API"
authors = ["brenopereira18 <brenopereira@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
factory-boy = "^3.3.1"
django = "^5.1.2"
django-rest-framework = "^0.1.0"
pytest-django = "^4.9.0"
djangorestframework = "^3.15.2"
django-extensions = "^3.2.3"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
